Split off Error veneer class into err.c/.h from veneer.c/.h
