+++ src/builtin/veneer.h
@@ -17,6 +17,7 @@
 #include "range.h"
 #include "strng.h"
 #include "dict.h"
+#include "err.h"
 
 #include "matrix.h"
 
@@ -24,20 +25,12 @@
  * Veneer classes
  * --------------------------- */
 
-#define ERROR_CLASSNAME "Error"
-
-#define ERROR_TAG_PROPERTY "tag"
-#define ERROR_MESSAGE_PROPERTY "message"
-
 #define ISMEMBER_ARG                      "IsMmbrArg"
 #define ISMEMBER_ARG_MSG                  "Method ismember expects a single argument."
 
 #define CLASS_INVK                        "ClssInvk"
 #define CLASS_INVK_MSG                    "Cannot invoke method '%s' on a class."
-
-#define ERROR_ARGS                        "ErrorArgs"
-#define ERROR_ARGS_MSG                    "Error much be called with a tag and a default message as arguments."
-
+ 
 void veneer_initialize(void);
 
 #endif /* veneer_h */
