+++ src/classes/range.c
@@ -7,7 +7,6 @@
 #include "morpho.h"
 #include "object.h"
 #include "builtin.h"
-#include "range.h"
 #include "veneer.h"
 
 /* **********************************************************************
